catch incorrectly sized objects at patch time instead of game runtime
