+++ structs/src/scly_props/thardus.rs
@@ -27,48 +27,12 @@ pub struct Thardus<'r>
     pub asset_ids2: GenericArray<u32, U6>,
 }
 
-use crate::{impl_position, impl_rotation, impl_scale, impl_patterned_info};
+use crate::{impl_position, impl_rotation, impl_scale, impl_patterned_info_with_auxillary};
 impl<'r> SclyPropertyData for Thardus<'r>
 {
     const OBJECT_TYPE: u8 = 0x58;
     impl_position!();
     impl_rotation!();
     impl_scale!();
-    impl_patterned_info!();
-
-    const SUPPORTS_DAMAGE_INFOS: bool = true;
-
-    fn impl_get_damage_infos(&self) -> Vec<DamageInfo> {
-        vec![
-            self.patterned_info.contact_damage.clone(),
-        ]
-    }
-
-    fn impl_set_damage_infos(&mut self, x: Vec<DamageInfo>) {
-        self.patterned_info.contact_damage = x[0].clone();
-    }
-
-    const SUPPORTS_VULNERABILITIES: bool = true;
-
-    fn impl_get_vulnerabilities(&self) -> Vec<DamageVulnerability> {
-        vec![
-            self.patterned_info.damage_vulnerability.clone(),
-        ]
-    }
-
-    fn impl_set_vulnerabilities(&mut self, x: Vec<DamageVulnerability>) {
-        self.patterned_info.damage_vulnerability = x[0].clone();
-    }
-
-    const SUPPORTS_HEALTH_INFOS: bool = true;
-
-    fn impl_get_health_infos(&self) -> Vec<HealthInfo> {
-        vec![
-            self.patterned_info.health_info.clone()
-        ]
-    }
-
-    fn impl_set_health_infos(&mut self, x: Vec<HealthInfo>) {
-        self.patterned_info.health_info = x[0].clone();
-    }
+    impl_patterned_info_with_auxillary!();
 }
